use arch-specific module for internal architecture dependent stuff
